génération des PlageHoraire, création de la classe PlageHoraire, correction Route

+++ src/modele/Route.java
@@ -6,11 +6,11 @@ public class Route {
 	private static int idRoutes =0;
 	
 	private String name;
-	private float vitesse,longueur;
+	private double vitesse,longueur;
 	private int id;
 	private Intersection inter;
 	
-	public Route(String nom, float vitesse, float longueur, Intersection inter1){
+	public Route(String nom, double vitesse, double longueur, Intersection inter1){
 		this.name = nom;
 		this.vitesse = vitesse;
 		this.longueur = longueur;
@@ -24,9 +24,9 @@ public class Route {
 	
 	public String getName(){return this.name;}
 	
-	public float getVitesse(){return this.vitesse;}
+	public double getVitesse(){return this.vitesse;}
 	
-	public float getLongueur(){return this.longueur;}
+	public double getLongueur(){return this.longueur;}
 	
 	public Intersection getInter(){return this.inter;}
